bug: omit session handler from serialization to avoid mp issues (#2366)
### Description
The session handler variable can be anything, because it's specific to
the SDK being used for the connector. This can break the serialization
depending on what that is. To avoid this all together, the session
handler itself is not serialized. Instead, it needs to be recreated if
an object is serialized and then deserialized.