rfctr: all ChunkingOptions are optional
Callers should not be in the business of applying default values for
chunking. Doing so as we have in the past leads to inconsistent
defaults, depending on whether chunking is applied by `unstructured`
locally, via ingest, or via one of the REST APIs or their SDK.

Make all chunking options `Optional[{type}]` such that a third-party
caller can simply use `None` for arguments omitted by the user and let
the chunker (via `ChunkingOptions`) be the sole arbiter of what default
value is applied.

Continue however to advertise certain default values (and use those
values for chunking defaults) so that UIs like the ingest CLI can
reliably advertise the default value even though they don't themselves
apply it.