rfctr: move combine_text_under_n_chars to by-title
So far, combining pre-chunks is a by-title-specific behavior. It has no
effect in the `basic` strategy because all pre-chunks are as big as
possible already. It really arose as a mitigation for mis-classification
of elements as titles so it only makes sense where titles represent a
semantic boundary.

However, it is a potentially general-purpose behavior and may be used in
forthcoming chunkers, but its default value and validation is likely to
differ.

Extract this option to `_ByTitleChunkingOptions` along with its
validation and tests. Leave a default in the `ChunkingOptions` base
class for use by subclasses but remove it from constructor and
validation (subclasses must validate).