rfctr: move boundary-predicates to opts
The now-differentiated options classes provide a better location for
defining boundary-predicates specific to that chunking strategy. Since
that was the only reason to subclass `BasePreChunker`, those subclasses
can now go away. Doing so further simplifies the objects needed to
create a pluggable chunking strategy.

Move the `.boundary_predicate` property to chunking-options and access
it from there.