step: add chunk()
This function adds a level of indirection so generic callers of chunkers
need not be coupled to (import) the set of chunkers available at
runtime. In particular this frees the `@add_chunking_strategy` decorator
from that responsibility.

Since pluggable-chunkers makes the available chunkers dynamic at
runtime, this is a critical requirement to enable pluggable chunkers.