feat(chunking): add metadata.orig_elements serde
This adds the needful such that `.metadata.orig_elements`, when present
on a chunk (element), is serialized to JSON when the chunk is
serialized, for instance, to be used in an HTTP response payload.

It also provides for "rehydrating" such a JSON payload into chunks that
contain the `.orig_elements` metadata.

Note that `.metadata.orig_elements` is always `Optional[list[Element]]`
when in memory. However, those original elements are serialized as
Base64-encoded gzipped JSON and are in that form (str) when present as
JSON or as "element-dicts" which is an interimediate
serialization/deserialization format. That is, serialization is `Element
-> dict -> JSON` and deserialization is `JSON -> dict -> `Element` and
`.orig_elements` are Base64-encoded in both the `dict` and `JSON` forms.