rfctr(xlsx): extract _XlsxPartitionerOptions
As an initial step in reducing the complexity of the monolithic
`partition_xlsx()` function, extract all argument-handling to a separate
`_XlsxPartitionerOptions` object which can be fully covered by isolated
unit tests.

This approach was incubated in the chunking context and has produced a
lot of opportunities to decompose the logic into smaller components that
are more understandable and isolated-test-able, without having to pass
an extended list of option values in ever sub-call. As well as
decluttering the code, this removes coupling where the caller needs to
know which options a subroutine might need to reference.