fix: `make pip-compile` should exit on failure
A `ResolutionImpossible` error during one of the many pip-compile
commands executed in series by `$ make pip-compile` does not cause the
`make` command to fail and gives the user no ready indication of the
failure.

Fix that by making the `foreach` loop used in `Makefile` exit on any
individual pip-compile command that fails.