fix(docx): fix disk-space leak
`partition_doc()` creates a temporary file in which it writes each
source-document provided as a file-like object. This file is not
deleted and disk consumption grows without bound.

`convert_office_doc()` uses a command-line program provided with
LibreOffice to convert from DOC -> DOCX. Because this command-line
program operates in a different memory space, the source file cannot be
passed as an in-memory object and needs to be on the filesystem.

Fix this by writing the temporary source file in the TemporaryDirectory
already being used to write the conversion-target DOCX file. That
directory is automatically removed when `partition_doc()` completes.